Allow unlocked USB data access (2/2)
by @KreAch3R

This bypasses the marshmallow "regression" of blocking and resetting USB data access
after disconnecting. It also allows saving the default configuration and it survives a reboot.
Originally, this "feature" was added by Google for security reasons
so it should be advised to use the below setting under caution.

Also:
 * Adds an "Advanced" section under Storage and USB
 * Moves Usb Configuration from Development settings
 * Secured lockscreen disables the option

Change-Id: I56d9ab1fa462e2ec74b28b3bc09c7934f760b37e