Settings: Match pulse speed to frameworks default
frameworks/base defines default LED timings:
 config_defaultNotificationLedOn = 500
 config_defaultNotificationLedOff = 2000

The LedOff timing of 2000 does not match any of the timings in the array
of notification_pulse_speed_values. So, a clean installation of
CyanogenMod results in an initial value of the notification pulse speed
being shown as "Custom". Reduce 2500ms in the pulse length and speed
arrays to 2000ms. Then, the default values for a new installation will
be: length = Short, speed = Slow.

This does have the side-effect of causing existing installations which
use 2500ms for either pulse length or speed now appearing as "Custom".

Change-Id: I8bbbb213c7b9645b0d62bfeb4620214ecf8dc4f5